refactor(web): promote V2 to canonical ForceGraph3D, retire V1
ADR-702's phase 1 plan: once V2 reaches parity, V1 is removed and the
unified-engine plugin becomes the one 3D force-graph. The action-layer
unification in PR #363 was the last gating step — V1 and V2 had been
routing identical mutations through useExplorationActions, so the V1
plugin was redundant.

Atomic because intermediate states (V1 deleted but routes still naming
'force-3d-v2', plugin renamed but registry still importing both) are
unshippable.

- Delete web/src/explorers/ForceGraph3D/ (V1 react-force-graph-3d
  wrapper, ~2050 lines)
- git mv ForceGraph3DV2/ → ForceGraph3D/; rename main file and
  V2SettingsPanel → SettingsPanel
- Rename internal symbols (ForceGraph3DV2 → ForceGraph3D,
  ForceGraph3DV2Data/Settings → ForceGraph3DData/Settings,
  V2SettingsPanel → SettingsPanel) and clean V1/V2 historical comments
- Drop 'force-3d-v2' from VisualizationType
- explorers/index.ts: stop exporting/importing ForceGraph3DV2Explorer
- App.tsx: drop the /explore/3d-v2 route; add a redirect from the old
  URL to /explore/3d so existing bookmarks resolve
- AppLayout.tsx sidebar: collapse two "3D Force Graph" entries into one
  (the Boxes icon stays, matching the new canonical entry)
- ExplorerView.tsx: drop the force-3d-v2 carve-out and the
  Settings3DPanel branch; the 3D plugin's own settingsPanel is used via
  the standard plugin contract. Only force-2d still routes through
  GraphSettingsPanel + SLIDER_RANGES_2D.
- Delete explorers/common/3DSettingsPanel.tsx (last consumer removed)
  and its export from common/index.ts
- Update on-canvas chrome label from "ForceGraph3D V2" to
  "ForceGraph3D"

Build, type-check, and the 10-test useExplorationActions suite all
clean against the renamed surface.
