+++ web/src/explorers/ForceGraph3D/scene/NodeInfoOverlay.tsx
