+++ web/src/explorers/ForceGraph3D/scene/gpuShaders.ts
