+++ web/src/explorers/ForceGraph3D/scene/useGpuForceSim.ts
