+++ web/src/explorers/common/index.ts
@@ -6,7 +6,6 @@
 export { NodeInfoBox, type NodeInfoBoxProps } from './NodeInfoBox';
 export { EdgeInfoBox, type EdgeInfoBoxProps } from './EdgeInfoBox';
 export { StatsPanel, type StatsPanelProps } from './StatsPanel';
-export { Settings3DPanel } from './3DSettingsPanel';
 export { GraphSettingsPanel } from './GraphSettingsPanel';
 export { Legend } from './Legend';
 export { PanelStack } from './PanelStack';
