+++ web/src/explorers/index.ts
@@ -7,11 +7,9 @@
 export * from './registry';
 export { ForceGraph2DExplorer } from './ForceGraph2D';
 export { ForceGraph3DExplorer } from './ForceGraph3D';
-export { ForceGraph3DV2Explorer } from './ForceGraph3DV2';
 export { DocumentExplorerPlugin } from './DocumentExplorer';
 
 // Import explorers to trigger auto-registration
 import './ForceGraph2D';
 import './ForceGraph3D';
-import './ForceGraph3DV2';
 import './DocumentExplorer';
